Phase 2.6 step 0 (D-02, DR-007): widen ListingColumnBatch.makeId to Int32Array
158 of the 295 makes in data/reference/taxonomy.json carry an identifier above
32 767 (max observed: 53 488), so the frozen Int16Array column wrapped them to
negative values: 158 out-of-taxonomy make cards carrying 17 988 of 100 000
listings on first paint, and makes unreachable in mode 2.

Amends the frozen 2.3 interface in both copies (docs/plans/DataProvider.ts and
src/providers/DataProvider.ts stay byte-for-byte identical), the column
descriptor, the synthetic columnar builder, the generator, the engine fixture
and the mock batches.

modelId keeps Int32Array: the highest model identifier in the taxonomy is
79 983, which the column already holds.

Probe made green without touching its assertion:
tests/review/D8/parcours.test.ts R-D8-31 (0 out-of-taxonomy cards on 295).

Probe adapted (D-31, type adaptation forced by the widening, justified by D-02):
tests/review/D2/dictionary-fields.test.ts asserted the pre-amendment physical
type of makeId; the three review fixtures that allocate the column themselves
were retyped as well.

+++ src/providers/DataProvider.ts
@@ -270,7 +270,7 @@ export interface ListingColumnBatch {
   readonly mileageKm: Int32Array;                   // sentinelle -1
   readonly firstRegistrationYearMonth: Int32Array;  // 12*année + (mois-1), sentinelle -1
   readonly modelId: Int32Array;                     // 0 = non résolu
-  readonly makeId: Int16Array;
+  readonly makeId: Int32Array;                      // ids AutoScout24 réels > 32 767 (D-02, DR-007)
   readonly modelYear: Int16Array;                   // sentinelle -1
   readonly powerKw: Int16Array;                     // sentinelle -1
   readonly co2EmissionsGPerKmX10: Int16Array;       // sentinelle -1
